+++ src/main/java/org/seedstack/scheduler/package-info.java
@@ -7,7 +7,5 @@
  */
 /**
  * Contains classes for scheduling tasks.
- *
- * @author Pierre Thirouin
  */
 package org.seedstack.scheduler;