+++ CHANGELOG.md
@@ -1,6 +1,11 @@
+# Version 3.0.0 (2017-01-03)
+
+* [new] Add a method to un-schedule a task by its trigger key. 
+* [brk] Update to new configuration system.
+
 # Version 2.2.0 (2016-08-10)
 
-* [new] Add the ability to substitue `@Scheduling` attributes `value`, `taskName` and `triggerName` with configuration values via `${}` placeholder.
+* [new] Add the ability to substitute `@Scheduling` attributes `value`, `taskName` and `triggerName` with configuration values via `${}` placeholder.
 
 # Version 2.1.1 (2016-04-26)